requirements.txt
appdirs==1.4.4
asgiref==3.2.7
attrs==19.3.0
black==19.10b0
boto3==1.13.23
botocore==1.16.23
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
codecov==2.1.4
coverage==5.1
defusedxml==0.6.0
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.7
django-allauth==0.42.0
django-crispy-forms==1.9.1
django-grappelli==2.14.2
django-recaptcha==2.0.6
djangorestframework==3.11.0
docutils==0.15.2
dynaconf==2.2.3
filelock==3.0.12
idna==2.9
isort==4.3.21
jmespath==0.10.0
oauthlib==3.1.0
pathspec==0.8.0
Pillow==7.1.2
pipenv==2020.6.2
psycopg2-binary==2.8.5
pudb==2019.2
Pygments==2.6.1
pyparsing==2.4.6
python-box==3.4.6
python-dateutil==2.8.1
python-dotenv==0.13.0
python3-openid==3.1.0
pytz==2020.1
PyYAML==5.3.1
regex==2020.5.14
requests==2.7.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
sentry-sdk==0.14.4
six==1.15.0
sqlparse==0.3.1
svgwrite==1.4
telebot==0.0.3
toml==0.10.1
Tree==0.2.4
typed-ast==1.4.1
urllib3==1.25.9
urwid==2.1.0
v==0.0.0
virtualenv==20.0.21
virtualenv-clone==0.5.4
whitenoise==5.1.0
